pyproject.toml
[project]
name = "sailer"
version = "0.1.0"
description = ""
authors = [
    {name = "Hailemichael Atrsaw Tibebu",email = "hailemichael.atrsaw2@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13.2, <4.0"
dependencies = [
    "dataconf (>=3.3.0,<4.0.0)",
    "croniter (>=6.0.0,<7.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "boto3 (>=1.37.3,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
